requirements.txt
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==0.0.6
django-debug-toolbar==1.9.1
django-disqus==0.5
django-tinymce==2.7.0
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.6
mock==2.0.0
pbr==3.1.1
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
stripe==1.79.1
urllib3==1.22
whitenoise==3.3.1
